requirements.txt
python-telegram-bot==20.7
Flask
python-dotenv
reportlab
